Fixed a bunch of compile errors with teh wasm64-wasi target

+++ library/std/src/sys/wasi/args.rs
@@ -21,7 +21,7 @@ fn maybe_args() -> Option<Vec<OsString>> {
     unsafe {
         let (argc, buf_size) = wasi::args_sizes_get().ok()?;
         let mut argv = Vec::with_capacity(argc);
-        let mut buf = Vec::with_capacity(buf_size);
+        let mut buf = Vec::with_capacity(buf_size as usize);
         wasi::args_get(argv.as_mut_ptr(), buf.as_mut_ptr()).ok()?;
         argv.set_len(argc);
         let mut ret = Vec::with_capacity(argc);
